% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dems_checks.R
\name{dems_check}
\alias{dems_check}
\title{Check whether one variable equals summation of other variables}
\usage{
dems_check(dems_df)
}
\arguments{
\item{dems_df}{Demographics Data Set}
}
\value{
A data set
}
\description{
Check whether one variable equals summation of other variables
}
