% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_negs.R
\name{no_negs}
\alias{no_negs}
\title{Checks that no negative values exist in data set}
\usage{
no_negs(latest_df)
}
\arguments{
\item{latest_df}{Any data frame}
}
\value{
A data frame.
}
\description{
Checks that no negative values exist in data set
}
