% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_checks.R
\name{only_date_cols}
\alias{only_date_cols}
\title{Checks vector/s from a data set to make sure that that vector/s has date values arranged as ymd}
\usage{
only_date_cols(df_date)
}
\arguments{
\item{df_date}{vector to be tested to the input file}

\item{date_order}{date order to be tested.}
}
\value{
tested vector
}
\description{
Checks vector/s from a data set to make sure that that vector/s has date values arranged as ymd
}
