% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threem_greater_sec.R
\name{threem_greater_sec}
\alias{threem_greater_sec}
\title{Checks whether variable from one data set equals values from another data set and returns NA for vals which do not meet condition}
\usage{
threem_greater_sec(df_1, df_2)
}
\arguments{
\item{df_1}{Data set 1}

\item{df_2}{Data set 2}
}
\value{
Data set
}
\description{
Checks whether variable from one data set equals values from another data set and returns NA for vals which do not meet condition
}
