% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_stop.R
\name{var_stop}
\alias{var_stop}
\title{Compare values in one variable between two data sets}
\usage{
var_stop(df_1, df_2, var_name)
}
\arguments{
\item{df_1}{Current data set.}

\item{df_2}{Previous data set.}

\item{var_name}{Variable name.}
}
\value{
data set.
}
\description{
Compare values in one variable between two data sets
}
